+++ src/main/java/com/kruemblegard/client/render/model/FaultCrawlerModel.java
@@ -5,9 +5,8 @@ import com.kruemblegard.entity.FaultCrawlerEntity;
 
 import net.minecraft.resources.ResourceLocation;
 
-import software.bernie.geckolib.model.GeoModel;
 
-public class FaultCrawlerModel extends GeoModel<FaultCrawlerEntity> {
+public class FaultCrawlerModel extends MirroredMobGeoModel<FaultCrawlerEntity> {
     @Override
     public ResourceLocation getModelResource(FaultCrawlerEntity animatable) {
         return new ResourceLocation(Kruemblegard.MOD_ID, "geo/fault_crawler.geo.json");