+++ src/main/java/com/kruemblegard/client/render/model/SpiralStriderModel.java
@@ -6,9 +6,8 @@ import com.kruemblegard.entity.adultform.SpiralStriderEntity;
 import net.minecraft.resources.ResourceLocation;
 
 import software.bernie.geckolib.core.animation.AnimationState;
-import software.bernie.geckolib.model.GeoModel;
 
-public class SpiralStriderModel extends GeoModel<SpiralStriderEntity> {
+public class SpiralStriderModel extends MirroredMobGeoModel<SpiralStriderEntity> {
     private static final int GOLEM_TEXTURE_VARIANT_MIN = 1;
     private static final int GOLEM_TEXTURE_VARIANT_MAX = 6;
 
@@ -38,6 +37,5 @@ public class SpiralStriderModel extends GeoModel<SpiralStriderEntity> {
     @Override
     public void setCustomAnimations(SpiralStriderEntity animatable, long instanceId, AnimationState<SpiralStriderEntity> animationState) {
         super.setCustomAnimations(animatable, instanceId, animationState);
-        AdultSleepPoseMirroring.apply(this, animatable);
     }
 }